add Ubuntu Mono font
I have been using this font on urxvt for months and it is really great.

This font was designed by Dalton Maag.

The .ttf files were copied from /usr/share/fonts/truetype/ubuntu-font-family/
directory.

I have tried it because Emacs GUI was using it and I thought I should give
it a try on the terminal.

It is also the default monospace font on Gnome:

> gsettings get org.gnome.desktop.interface monospace-font-name
'Ubuntu Mono 13'

Refs:
https://www.fontsquirrel.com/fonts/ubuntu-mono

Signed-off-by: Thiago Farina <tfransosi@gmail.com>