add info about fonts used in Ubuntu 16.04.7 with Xfce 4.12
On Xfce, the default font can be located at Appearance dialog.
To open it, click on Start menu > Settings > Appearance > and select
the Fonts tab.

The Window titlebar font, is located at the Window Manager dialog.
To open it, click on Start menu > Settings > Window Manager > and
select the Style tab.

The keys were found with xfconf-query 4.12.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>